+++ Tests/LinuxMain.swift
@@ -266,6 +266,14 @@ extension MultilineArgumentsRuleTests {
     ]
 }
 
+extension ModifiersOrderTests {
+    static var allTests: [(String, (ModifiersOrderTests) -> () throws -> Void)] = [
+        ("testAttibuteStatic", testAttibuteStatic),
+        ("testAttibuteClass", testAttibuteClass),
+        ("testAttributesOrderLeft", testAttributesOrderLeft)
+    ]
+}
+
 extension NumberSeparatorRuleTests {
     static var allTests: [(String, (NumberSeparatorRuleTests) -> () throws -> Void)] = [
         ("testNumberSeparatorWithDefaultConfiguration", testNumberSeparatorWithDefaultConfiguration),
@@ -427,6 +435,7 @@ extension RulesTests {
         ("testLegacyConstructor", testLegacyConstructor),
         ("testLetVarWhitespace", testLetVarWhitespace),
         ("testLiteralExpressionEndIdentation", testLiteralExpressionEndIdentation),
+        ("testModifiersOrderRule", testModifiersOrderRule),
         ("testMark", testMark),
         ("testMultilineParameters", testMultilineParameters),
         ("testMultipleClosuresWithTrailingClosure", testMultipleClosuresWithTrailingClosure),
@@ -577,6 +586,7 @@ XCTMain([
     testCase(LineLengthRuleTests.allTests),
     testCase(LinterCacheTests.allTests),
     testCase(MultilineArgumentsRuleTests.allTests),
+    testCase(ModifiersOrderTests.allTests),
     testCase(NumberSeparatorRuleTests.allTests),
     testCase(ObjectLiteralRuleTests.allTests),
     testCase(PrivateOverFilePrivateRuleTests.allTests),